♻️ move types to a dedicated file
