+++ packages/rum/src/index.ts
@@ -1,4 +1,10 @@
 export { Datacenter } from '@datadog/browser-core'
-export { RumUserConfiguration, RumGlobal, datadogRum, InternalContext } from './rum.entry'
-export { RumEventCategory, RumResourceEvent, RumViewEvent, RumUserActionEvent } from './rum'
-export { RumEvent } from './assembly'
+export { RumUserConfiguration, RumGlobal, datadogRum } from './rum.entry'
+export {
+  InternalContext,
+  RumEvent,
+  RumEventCategory,
+  RumUserActionEvent,
+  RumViewEvent,
+  RumResourceEvent,
+} from './types'
