+++ packages/rum/src/rum.entry.ts
@@ -3,7 +3,6 @@ import {
   buildCookieOptions,
   checkCookiesAuthorized,
   checkIsNotLocalFile,
-  combine,
   Context,
   createContextManager,
   deepClone,
@@ -20,19 +19,6 @@ import { CustomUserAction, UserActionType } from './userActionCollection'
 
 export interface RumUserConfiguration extends UserConfiguration {
   applicationId: string
-}
-
-export interface InternalContext {
-  application_id: string
-  session_id: string | undefined
-  view?: {
-    id: string
-    url: string
-    referrer: string
-  }
-  user_action?: {
-    id: string
-  }
 }
 
 export type RumGlobal = ReturnType<typeof makeRumGlobal>
