+++ packages/rum/src/rum.ts
@@ -3,11 +3,9 @@ import {
   commonInit,
   Configuration,
   Context,
-  ErrorContext,
   ErrorMessage,
   generateUUID,
   getTimestamp,
-  HttpContext,
   includes,
   msToNs,
   RequestType,
@@ -31,107 +29,20 @@ import {
   computeResourceKind,
   computeSize,
 } from './resourceUtils'
-import { InternalContext, RumUserConfiguration } from './rum.entry'
+import { RumUserConfiguration } from './rum.entry'
 import { RumSession, startRumSession } from './rumSession'
-import { CustomUserAction, startUserActionCollection, UserActionMeasures, UserActionType } from './userActionCollection'
-import { startViewCollection, ViewLoadingType, ViewMeasures } from './viewCollection'
-
-export enum RumEventCategory {
-  USER_ACTION = 'user_action',
-  ERROR = 'error',
-  LONG_TASK = 'long_task',
-  VIEW = 'view',
-  RESOURCE = 'resource',
-}
-
-interface PerformanceResourceDetailsElement {
-  duration: number
-  start: number
-}
-
-export interface PerformanceResourceDetails {
-  redirect?: PerformanceResourceDetailsElement
-  dns?: PerformanceResourceDetailsElement
-  connect?: PerformanceResourceDetailsElement
-  ssl?: PerformanceResourceDetailsElement
-  firstByte: PerformanceResourceDetailsElement
-  download: PerformanceResourceDetailsElement
-}
-
-export interface RumResourceEvent {
-  date: number
-  duration: number
-  evt: {
-    category: RumEventCategory.RESOURCE
-  }
-  http: {
-    performance?: PerformanceResourceDetails
-    method?: string
-    statusCode?: number
-    url: string
-  }
-  network: {
-    bytesWritten?: number
-  }
-  resource: {
-    kind: ResourceKind
-    id?: string // only for traced requests
-  }
-  _dd?: {
-    traceId: string
-    spanId?: string // not available for initial document tracing
-  }
-}
-
-export interface RumErrorEvent {
-  date: number
-  http?: HttpContext
-  error: ErrorContext
-  evt: {
-    category: RumEventCategory.ERROR
-  }
-  message: string
-}
-
-export interface RumViewEvent {
-  date: number
-  duration: number
-  evt: {
-    category: RumEventCategory.VIEW
-  }
-  rum: {
-    documentVersion: number
-  }
-  view: {
-    loadingTime?: number
-    loadingType: ViewLoadingType
-    measures: ViewMeasures
-  }
-}
-
-export interface RumLongTaskEvent {
-  date: number
-  duration: number
-  evt: {
-    category: RumEventCategory.LONG_TASK
-  }
-}
-
-export interface RumUserActionEvent {
-  date?: number
-  duration?: number
-  evt: {
-    category: RumEventCategory.USER_ACTION
-    name: string
-  }
-  userAction: {
-    id?: string
-    type: UserActionType
-    measures?: UserActionMeasures
-  }
-}
-
-export type RawRumEvent = RumErrorEvent | RumResourceEvent | RumViewEvent | RumLongTaskEvent | RumUserActionEvent
+import {
+  InternalContext,
+  RawRumEvent,
+  RumErrorEvent,
+  RumEventCategory,
+  RumLongTaskEvent,
+  RumResourceEvent,
+  RumUserActionEvent,
+  RumViewEvent,
+} from './types'
+import { CustomUserAction, startUserActionCollection } from './userActionCollection'
+import { startViewCollection } from './viewCollection'
 
 export function startRum(userConfiguration: RumUserConfiguration, getGlobalContext: () => Context) {
   const lifeCycle = new LifeCycle()
