+++ packages/rum/test/specHelper.ts
@@ -12,8 +12,9 @@ import { startRumAssembly } from '../src/assembly'
 import { LifeCycle, LifeCycleEventType } from '../src/lifeCycle'
 import { ParentContexts, startParentContexts } from '../src/parentContexts'
 import { startPerformanceCollection } from '../src/performanceCollection'
-import { RawRumEvent, startRumEventCollection } from '../src/rum'
+import { startRumEventCollection } from '../src/rum'
 import { RumSession } from '../src/rumSession'
+import { RawRumEvent } from '../src/types'
 import { startUserActionCollection } from '../src/userActionCollection'
 import { startViewCollection } from '../src/viewCollection'
